requirements.txt
pycdlib>=1.14.0
capstone>=5.0
unicorn>=2.0
openai>=1.0
anthropic>=0.20
pyyaml>=6.0
python-dotenv>=1.0
chardet>=5.0
click>=8.1
rich>=13.0
